Some optimization

+++ Utils.h
@@ -16,7 +16,7 @@ public:
   ~LinearCombination();
   std::string Key();
   uint8_t nodeId;
-  uint8_t index;
+  uint16_t index;
 	uint8_t coeff;
 	uint8_t dstId;
 	uint32_t genTime;
@@ -42,7 +42,7 @@ class NCAttribute : public ns3::Object {
 	// data members
 public:
 	uint8_t m_nodeId;
-	uint8_t m_index;
+	uint16_t m_index;
 	uint8_t m_destId;
 	uint32_t m_genTime;
 	uint32_t m_receptionNum;
@@ -51,13 +51,13 @@ public:
 	bool m_destReceived;
 	NCAttribute();
 	NCAttribute(const NCAttribute& );
-	NCAttribute (const uint8_t nodeId,const int8_t index,const uint8_t id, const uint32_t genTime);
+	NCAttribute (const uint8_t nodeId,const int16_t index,const uint8_t id, const uint32_t genTime);
 	~NCAttribute ();
 	//member functions
 	void SetNodeId (uint8_t nodeId);
 	uint8_t GetNodeId () const;
-	void SetIndex (uint8_t index);
-	uint8_t GetIndex () const;
+	void SetIndex (uint16_t index);
+	uint16_t GetIndex () const;
 	void SetGenTime (uint32_t genTime);
 	uint32_t GetGenTime () const;
 	void SetDestination (const uint8_t ip);
@@ -74,20 +74,20 @@ class CoefElt
 public:
 	// Data Members
 	uint8_t m_coef;
-	uint8_t m_index;
+	uint16_t m_index;
 	uint8_t m_nodeId;
 	uint8_t m_destId;
 	uint8_t m_length;
 	uint32_t m_genTime;
 	// Constructors and a Destructor
 	CoefElt ();
-	CoefElt (uint8_t coef, uint8_t index, uint8_t nodeId, uint32_t genTime);
+	CoefElt (uint8_t coef, uint16_t index, uint8_t nodeId, uint32_t genTime);
 	~CoefElt ();
 	// Methods
 	uint8_t GetCoef () const;
 	void SetCoef (uint8_t);
-	uint8_t GetIndex() const;
-	void SetIndex (uint8_t index);
+	uint16_t GetIndex() const;
+	void SetIndex (uint16_t index);
 	uint8_t GetNodeId ();
 	void SetNodeId (uint8_t nodeId);
 	uint32_t GetGenTime () const;
@@ -110,7 +110,7 @@ public:
 	//galois::GaloisField *m_galoisField;
 	MapType m_coefsList;
 	int m_dataLength;
-	uint8_t m_index;
+	uint16_t m_index;
 	bool m_decoded;
 	//char * payload;
 
@@ -123,8 +123,8 @@ public:
 	bool operator== (const NetworkCodedDatagram& nc) const;
 	// Methods
 	// void InitializeGaloisField ();
-	uint8_t GetIndex () const;
-	void SetIndex (uint8_t index);
+	uint16_t GetIndex () const;
+	void SetIndex (uint16_t index);
 	bool IsNull ();
 	int GetLength() const;
 	void SetLength (int length);
